renamed aws resouces to follow naming conventions (#49)
* renamed aws resouces to follow naming conventions

hosted zone was fine as is
vpc -> vpc instance
compute -> compute engine
storage -> bucket
