+++ examples/particle/aws/vpc/use_security_group/example_security_group_with_vpc_parent.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from pcf.particle.aws.vpc.security_group import SecurityGroup
-from pcf.particle.aws.vpc.vpc import VPC
+from pcf.particle.aws.vpc.vpc_instance import VPC
 from pcf.core.quasiparticle import Quasiparticle
 from pcf.core import State
