+++ examples/particle/aws/vpc/use_subnet/example_subnet_with_vpc/example_subnet_with_vpc.py
@@ -1,5 +1,5 @@
 from pcf.particle.aws.vpc.subnet import Subnet
-from pcf.particle.aws.vpc.vpc import VPC
+from pcf.particle.aws.vpc.vpc_instance import VPC
 from pcf.core.quasiparticle import Quasiparticle
 from pcf.core import State
