+++ pcf/test/particle/aws/vpc/test_vpc_instance.py
@@ -14,7 +14,7 @@
 
 import moto
 
-from pcf.particle.aws.vpc.vpc import VPC
+from pcf.particle.aws.vpc.vpc_instance import VPC
 from pcf.core import State
